requirements.txt
streamlit
pandas
Pillow
confluent-kafka
